+++ CHANGELOG.md
@@ -5,6 +5,32 @@ All notable changes to `Laragram` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.0.2] (2019-02-10)
+
+### Added
+
+- Add a package documentation to the [Wiki](https://github.com/wekser/laragram/wiki)
+- Add possibility of returning a string from a route or controller as easy text message
+- Add `ResponseInvalidException`
+
+### Changed
+
+- Replaced `get` on `input` and `input` on `query` methods in `BotRequest`
+
+### Fixed
+
+- Fixed and update PHPDoc descriptions
+
+## [v1.0.1] (2019-02-08)
+
+### Added
+
+- Add Extensions
+
+### Changed
+
+- Update work of adding route in collection
+
 ## [v1.0.0] (2019-02-05)
 
 - This initial release